Move OwnersClient config to plugins config.
